feat(contracts): adding price deterministic functionality for DEX identification
summary: the core algorithm which identifies a DEX to execute a trade against has thus far ben designed so as to observe the DEX with the highest reserve depth - this is done so as to ensure price impact is always minimal. However, a user may wish to override this and instead utilise a price based approach, executing against a DEX which has the lowest price for the token pair. It's a risky approach, since whilst a price may be attractive, the price impact with a DEX that has low reserves would be unpredictable, resulting in what may be an enormous sweet spot and just an unpragmatic financial solution. However, for the sake of future proofing and flexibility, this option has now been integrated. A flag is included in a trade struct which detemrines wether or not to use price or reserve determinism when executing a trade. Surrounding test files have accordingly been updated, although no explicit price based tests exist thus far. It is not reccomended, therefore, at this stage to utilise this feature until due testing has been undertaken.